refactor body UI code
