+++ selfdrive/ui/body/layouts/__init__.py
@@ -1,3 +1,3 @@
-from openpilot.selfdrive.ui.body.layouts.body import BodyLayout
-from openpilot.selfdrive.ui.body.layouts.body_sidebar import BodySidebar, BODY_SIDEBAR_HEIGHT
-from openpilot.selfdrive.ui.body.layouts.home import MiciBodyHomeLayout
+from openpilot.selfdrive.ui.body.layouts.home_tici import BodyLayout
+from openpilot.selfdrive.ui.body.widgets.sidebar_tici import BodySidebar, BODY_SIDEBAR_HEIGHT
+from openpilot.selfdrive.ui.body.layouts.home_mici import MiciBodyHomeLayout
