+++ selfdrive/ui/body/widgets/sidebar_tici.py
